requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
asgiref==3.8.1
    # via django
django==4.2.16
    # via
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
    #   wingz-app (setup.py)
django-extensions==3.2.3
    # via wingz-app (setup.py)
django-filter==24.3
    # via wingz-app (setup.py)
djangorestframework==3.15.2
    # via
    #   djangorestframework-simplejwt
    #   drf-yasg
    #   wingz-app (setup.py)
djangorestframework-simplejwt==5.3.1
    # via wingz-app (setup.py)
drf-yasg==1.21.8
    # via wingz-app (setup.py)
inflection==0.5.1
    # via drf-yasg
packaging==24.1
    # via drf-yasg
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pytz==2024.2
    # via drf-yasg
pyyaml==6.0.2
    # via drf-yasg
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via asgiref
uritemplate==4.1.1
    # via drf-yasg
